+++ src/common/Collision/RegularGrid.h
@@ -18,8 +18,8 @@
 #ifndef _REGULAR_GRID_H
 #define _REGULAR_GRID_H
 
-#include "Containers.h"
 #include "Errors.h"
+#include "IteratorPair.h"
 #include <G3D/Ray.h>
 #include <G3D/BoundsTrait.h>
 #include <G3D/PositionTrait.h>
