+++ src/server/game/Combat/ThreatManager.h
@@ -20,11 +20,11 @@
 #define _THREATMANAGER
 
 #include "Common.h"
+#include "IteratorPair.h"
 #include "SharedDefines.h"
 #include "LinkedReference/Reference.h"
 #include "UnitEvents.h"
 #include "ObjectGuid.h"
-#include "Containers.h"
 
 #include <list>
 
@@ -220,8 +220,8 @@ class TC_GAME_API ThreatManager
 {
     public:
         // -- compatibility layer for combat rewrite (PR #19930)
-        Trinity::Containers::IteratorPair<std::list<ThreatReference*>::const_iterator> GetSortedThreatList() const { auto& list = iThreatContainer.getThreatList(); return { list.cbegin(), list.cend() }; }
-        Trinity::Containers::IteratorPair<std::list<ThreatReference*>::const_iterator> GetUnsortedThreatList() const { return GetSortedThreatList(); }
+        Trinity::IteratorPair<std::list<ThreatReference*>::const_iterator> GetSortedThreatList() const { auto& list = iThreatContainer.getThreatList(); return { list.cbegin(), list.cend() }; }
+        Trinity::IteratorPair<std::list<ThreatReference*>::const_iterator> GetUnsortedThreatList() const { return GetSortedThreatList(); }
         std::list<ThreatReference*> GetModifiableThreatList() const { return iThreatContainer.getThreatList(); }
         Unit* SelectVictim() { return getHostilTarget(); }
         Unit* GetCurrentVictim() const { if (ThreatReference* ref = getCurrentVictim()) return ref->GetVictim(); else return nullptr; }
